+++ src/main/java/CloudProvider/AWS/AwsCloudClient.java
@@ -1,5 +1,8 @@
 package CloudProvider.AWS;
 
+import java.util.List;
+import java.util.Map;
+
 import CloudProvider.ICloudClient;
 import software.amazon.awssdk.auth.credentials.ProfileCredentialsProvider;
 
@@ -12,7 +15,7 @@ public class AwsCloudClient implements ICloudClient {
 
     private AwsCloudClient(){
         ProfileCredentialsProvider profile = ProfileCredentialsProvider.create();
-        objectHelper = new AwsDataObjectHelper(profile);
+        objectHelper = new AwsDataObjectHelper(profile, null);
         labelHelper = new AwsLabelDetectorHelper(profile);
     }
 
@@ -28,5 +31,16 @@ public class AwsCloudClient implements ICloudClient {
         labelHelper.Close();
         INSTANCE = null;
     }
-    
+
+    public void CreateObject(String objectName, String objectPath){
+        objectHelper.CreateObject(objectName, objectPath);
+    }
+
+    public void CreateBucket(String bucket){
+        objectHelper.CreateBucket(bucket);
+    }
+
+    public List<String> Execute(String imageUri, Map<String, Object> params){
+        return labelHelper.Execute(imageUri, params);
+    }
 }
