+++ src/main/java/CloudProvider/ILabelDetector.java
@@ -1,5 +1,8 @@
 package CloudProvider;
 
+import java.util.List;
+import java.util.Map;
+
 public interface ILabelDetector {
-    public String Execute(String imageUri, int[] params);
+    public List<String> Execute(String imageUri, Map<String, Object> params);
 }
